# WAIL OUAISSA GraphQL project 2/WAIL OUAISSA GraphQL Server/src/main/java/com/example/DemoGraphQL/model/Comment.java
package com.example.DemoGraphQL.model;

import lombok.Data;

import javax.persistence.*;
import java.time.LocalDateTime;
import java.util.Date;

@Entity
@Table(name = "comments")
@Data
public class Comment {

    @Id
    @GeneratedValue(strategy = GenerationType.AUTO)
    private Long id;

    private String content;

    @ManyToOne
    @JoinColumn(name = "user_id")
    private User user;

    @ManyToOne
    @JoinColumn(name = "post_id")
    private Post post;

    private Date creationDate;


}